improves user experience on mobile devices
