+++ src/components/molecules/MobileNavitagion/MobileNavitagion.stories.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import { storiesOf } from '@storybook/react'
+
+import MobileNavitagion from './MobileNavitagion'
+
+storiesOf('MobileNavitagion', module).add('renders a mobile navigation', () => (
+  <MobileNavitagion />
+))
